Fast plugins can stay in "kotlin-jvm"
