+++ kotlin-jvm/src/main/graphql/launches/schema.sdl
