+++ kotlin-jvm/src/main/kotlin/playground/KotlinxSerialization.kt
@@ -56,3 +56,11 @@ internal data class Users(
     val size: Int,
     val users: List<User> = emptyList()
 )
+
+@Serializable
+data class HttpbinGet(
+    val args: Map<String, String> = emptyMap(),
+    val headers: Map<String, String> = emptyMap(),
+    val origin: String,
+    val url: String
+)
