+++ kotlin-jvm/src/main/kotlin/playground/Retrofit.kt
@@ -8,7 +8,7 @@ import kotlinx.serialization.json.Json
 import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
-import playground.HttpbinGet
+import playground.kotlinx.serialization.HttpbinGet
 import playground.shouldBe
 import retrofit2.Call
 import retrofit2.Retrofit
