+++ kotlin-jvm/src/main/sqldelight/util/user.sq
